requirements.txt
attrs==22.1.0
black==22.8.0
click==8.1.3
Flask==2.2.2
importlib-metadata==4.12.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mypy-extensions==0.4.3
packaging==21.3
passlib==1.7.4
pathspec==0.10.1
platformdirs==2.5.2
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
PyJWT==1.4.2
pyparsing==3.0.9
pytest==7.1.3
python-dotenv==0.21.0
tomli==2.0.1
typing-extensions==4.3.0
Werkzeug==2.2.2
zipp==3.8.1
